basic sonification, but needs work

+++ nars_gui/src/main/java/automenta/vivisect/audio/granular/GrainWindow.java
@@ -12,12 +12,12 @@ public class GrainWindow {
 		return factors.length;
 	}
 
-	public float getFactor(long offset) {
-		long index = offset;
+	public float getFactor(int offset) {
+		int index = offset;
 		if (offset < 0) { // Fade in
 			index = -offset;
 		}
-		return index < factors.length ? factors[(int)index] : 0.0F;
+		return index < factors.length ? factors[index] : 0.0F;
 	}
 
 }
